Книга рецептов-продолжение-06.03.23

+++ src/main/java/me/ruana/recipeBook/dto/IngredientsDTO.java
@@ -6,11 +6,11 @@ import java.util.Objects;
 
 public class IngredientsDTO {
     private final String ingredientName;
-    private final int numberOfIngredients;
+    private final double numberOfIngredients;
     private final String measureUnits;
     private final int id;
 
-    public IngredientsDTO (int id, String ingredientName, int numberOfIngredients, String measureUnits) {
+    public IngredientsDTO(int id, String ingredientName, double numberOfIngredients, String measureUnits) {
         this.ingredientName = ingredientName;
         this.numberOfIngredients = numberOfIngredients;
         this.measureUnits = measureUnits;
@@ -21,7 +21,7 @@ public class IngredientsDTO {
         return ingredientName;
     }
 
-    public int getNumberOfIngredients() {
+    public double getNumberOfIngredients() {
         return numberOfIngredients;
     }
 
@@ -44,13 +44,14 @@ public class IngredientsDTO {
     public int hashCode() {
         return Objects.hash(getIngredientName(), getNumberOfIngredients(), getMeasureUnits(), getId());
     }
-    public static IngredientsDTO from(int id, Ingredients ingredients){
+
+    public static IngredientsDTO from(int id, Ingredients ingredients) {
         return new IngredientsDTO(id, ingredients.getIngredientName(), ingredients.getNumberOfIngredients(), ingredients.getMeasureUnits());
     }
 
     @Override
     public String toString() {
-        return  id + ". " + "Наименование: " + ingredientName + '\'' +
+        return id + ". " + "Наименование: " + ingredientName + '\'' +
                 "\n Количество " + numberOfIngredients +
                 " " + measureUnits;
     }
